Add generic action-cancellation mechanism; range fields are now int32
Two related changes requested together:

Range as int32, not float. min_range/max_range (attack_resolution,
auto_attack, cast_time_attack) and AuraSource/ActivateAura's range are
authored content - a weapon's or spell's range is a round number ("5
yards"), never a computed continuous quantity. movement::Position and
the distances derived from it stay float; only the configured
threshold being compared against is now int32, cast to float once at
each comparison site.

Generic cancellation mechanism, answering "do we really need a
separate mechanism per capability for this?" - no: a new interruption
capability provides ActionInterrupted{entity}, a deliberately
capability-agnostic event (no properties, no requests, no
hand-written .cpp at all - the first capability in this demo that's
purely a shared vocabulary event). Two triggers:

- Movement, opt-in per attack (WeaponAttack/CastTimeAttack both gain a
  requires_stationary flag). Nothing about movement::on_move changed -
  it already published PositionChanged; cast_time_attack and
  auto_attack simply subscribe to that existing event and check their
  own flag before reacting, so an attack that never opted in is
  untouched by movement.
- interruption::ActionInterrupted, unconditional - the actual generic
  piece, for a future crowd-control effect (stun, disorient) to
  publish. Both capabilities subscribe and cancel regardless of
  requires_stationary.

Each capability decides for itself what "cancelled" means (spec §20's
Design Rule: never reach into another capability's state):
cast_time_attack cancels outright to a clean idle state; auto_attack,
which has no in-progress wind-up to cancel, instead resets its
cooldown to a full cycle whenever a swing was mid-countdown - already-
ready weapons are left alone in both cases, since there's nothing in
progress to interrupt.

SimulatedHost's constructor wires both capabilities' handlers against
movement::PositionChanged and interruption::ActionInterrupted - host
composition deciding which capabilities react to which events, the
same kind of decision registering property stores already is.

No crowd-control capability is built here - the mechanism is proven by
publishing ActionInterrupted directly from tests, the same
prove-the-shape-without-the-whole-feature precedent
InstantAttackBypassesTheAutoAttackCooldownEntirely already set.

+++ demo/modules/attack_resolution/attack_resolution.cpp
@@ -23,7 +23,7 @@ TargetedAttackOutcome resolve_targeted_attack(Context& ctx, const TargetedAttack
     const float delta_y = target_position->get().y - attacker_position->get().y;
     const float distance = std::sqrt((delta_x * delta_x) + (delta_y * delta_y));
 
-    if (distance < query.min_range || distance > query.max_range) {
+    if (distance < static_cast<float>(query.min_range) || distance > static_cast<float>(query.max_range)) {
         return TargetedAttackOutcome{.result = accept(query), .landed = false};
     }
